Add FlashInfer-style transformer kernel inventory
