+++ src/tilegym/transformers/olmo3/modeling_olmo3.py
@@ -4,123 +4,11 @@
 
 """OLMo-3-specific cuTile kernel wrappers and patched decoder layer forward."""
 
-import cuda.tile as ct
 import torch
 import torch.nn as nn
 
-ConstInt = ct.Constant[int]
-
-
-@ct.kernel
-def _dual_rms_norm_kernel(
-    q,  # (N, D) — projected Q, normalized in-place
-    k,  # (N, D) — projected K, normalized in-place
-    q_weight,  # (D,)
-    k_weight,  # (D,)
-    eps: float,
-    D: ConstInt,
-    TILE_D: ConstInt,
-):
-    """Fused in-place RMSNorm for Q and K in a single kernel launch.
-
-    Grid: (N,). Each block normalizes both q[bid] and k[bid] in-place.
-    No race conditions since each block owns a unique row pair.
-    """
-    PAD = ct.PaddingMode.ZERO
-    bid = ct.bid(0)
-
-    # ---- Normalize Q row ----
-    q_h = ct.load(q, index=(bid, 0), shape=(1, TILE_D), padding_mode=PAD).reshape((TILE_D,)).astype(ct.float32)
-    q_w = ct.load(q_weight, index=(0,), shape=(TILE_D,), padding_mode=PAD).astype(ct.float32)
-    q_var = ct.sum(q_h * q_h) * ct.truediv(1.0, D)
-    q_normed = q_h * ct.rsqrt(q_var + eps) * q_w
-    ct.store(q, index=(bid, 0), tile=q_normed.reshape((1, TILE_D)).astype(q.dtype))
-
-    # ---- Normalize K row ----
-    k_h = ct.load(k, index=(bid, 0), shape=(1, TILE_D), padding_mode=PAD).reshape((TILE_D,)).astype(ct.float32)
-    k_w = ct.load(k_weight, index=(0,), shape=(TILE_D,), padding_mode=PAD).astype(ct.float32)
-    k_var = ct.sum(k_h * k_h) * ct.truediv(1.0, D)
-    k_normed = k_h * ct.rsqrt(k_var + eps) * k_w
-    ct.store(k, index=(bid, 0), tile=k_normed.reshape((1, TILE_D)).astype(k.dtype))
-
-
-def dual_rms_norm_cutile(
-    q: torch.Tensor,
-    k: torch.Tensor,
-    q_weight: torch.Tensor,
-    k_weight: torch.Tensor,
-    eps: float,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Fused in-place RMSNorm for Q and K in a single kernel launch."""
-    D = q.shape[-1]
-    q_flat = q.contiguous().view(-1, D)
-    k_flat = k.contiguous().view(-1, D)
-    N = q_flat.shape[0]
-    TILE_D = 1 << (D - 1).bit_length()
-    ct.launch(
-        torch.cuda.current_stream(),
-        (N,),
-        _dual_rms_norm_kernel,
-        (q_flat, k_flat, q_weight, k_weight, eps, D, TILE_D),
-    )
-    return q, k
-
-
-@ct.kernel
-def _rms_norm_residual_add_kernel(
-    x,  # (N, D) — branch output (attn or mlp)
-    residual,  # (N, D) — residual from before the branch
-    weight,  # (D,) — RMSNorm weight
-    out,  # (N, D) — residual + rms_norm(x)
-    eps: float,
-    D: ConstInt,
-    TILE_D: ConstInt,
-):
-    """Fused RMSNorm + residual add for OLMo-3 post-normalization.
-
-    Computes: out = residual + weight * x * rsqrt(mean(x^2) + eps)
-
-    OLMo-3 uses post-norm (norm on branch output, then add residual),
-    so this fuses what would be two separate ops into one kernel.
-    """
-    bid = ct.bid(0)
-    offs = ct.arange(TILE_D, dtype=ct.int32)
-
-    # Load in float32 for numerical stability
-    h = ct.astype(ct.gather(x, (bid, offs), padding_value=0.0, check_bounds=True), ct.float32)
-    r = ct.astype(ct.gather(residual, (bid, offs), padding_value=0.0, check_bounds=True), ct.float32)
-    w = ct.astype(ct.gather(weight, (offs,), padding_value=0.0, check_bounds=True), ct.float32)
-
-    # RMSNorm: weight * x * rsqrt(mean(x^2) + eps)
-    variance = ct.sum(h * h) * ct.truediv(1.0, D)
-    normed = h * ct.rsqrt(variance + eps) * w
-
-    # Residual add
-    result = r + normed
-
-    ct.scatter(out, (bid, offs), ct.astype(result, out.dtype), check_bounds=True)
-
-
-def rms_norm_residual_add_cutile(
-    x: torch.Tensor,
-    residual: torch.Tensor,
-    weight: torch.Tensor,
-    eps: float,
-) -> torch.Tensor:
-    """Fused RMSNorm + residual add. Returns residual + rms_norm(x)."""
-    D = x.shape[-1]
-    x_flat = x.contiguous().view(-1, D)
-    r_flat = residual.contiguous().view(-1, D)
-    N = x_flat.shape[0]
-    out = torch.empty_like(x_flat)
-    TILE_D = 1 << (D - 1).bit_length()
-    ct.launch(
-        torch.cuda.current_stream(),
-        (N,),
-        _rms_norm_residual_add_kernel,
-        (x_flat, r_flat, weight, out, eps, D, TILE_D),
-    )
-    return out.view(x.shape)
+from tilegym.transformers.olmo3.kernels.dual_rms_norm import dual_rms_norm_cutile
+from tilegym.transformers.olmo3.kernels.rms_norm_residual_add import rms_norm_residual_add_cutile
 
 
 class FusedOlmo3MLP(nn.Module):
