+++ src/tilegym/transformers/olmoe/kernel_definitions/olmoe_residual_add_rms_norm.json
@@ -1,0 +1,65 @@
+{
+  "axes": {
+    "D": {
+      "description": "Normalized dimension.",
+      "type": "var"
+    },
+    "N": {
+      "description": "Flattened leading dimensions.",
+      "type": "var"
+    }
+  },
+  "description": "OLMoE fused residual add and Llama-style RMSNorm returning both sum and normalized output.",
+  "inputs": {
+    "eps": {
+      "dtype": "float32",
+      "shape": null
+    },
+    "residual": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    },
+    "weight": {
+      "dtype": "bfloat16",
+      "shape": [
+        "D"
+      ]
+    },
+    "x": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    }
+  },
+  "name": "olmoe_residual_add_rms_norm",
+  "op_type": "fused_add_rmsnorm",
+  "outputs": {
+    "normed": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    },
+    "sum": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    }
+  },
+  "reference": "# Source: https://github.com/huggingface/transformers/blob/aad13b87ed59f2afcfaebc985f403301887a35fc/src/transformers/models/olmoe/modeling_olmoe.py#L414-L419\n# Source: https://github.com/huggingface/transformers/blob/aad13b87ed59f2afcfaebc985f403301887a35fc/src/transformers/models/olmoe/modeling_olmoe.py#L58-L63\nimport torch\n\ndef run(eps, residual, weight, x):\n    summed = residual + x\n    s = summed.to(torch.float32)\n    normed = s * torch.rsqrt((s * s).mean(-1, keepdim=True) + eps) * weight\n    return normed.to(summed.dtype), summed",
+  "tags": [
+    "model:olmoe",
+    "stage:prefill",
+    "stage:decode",
+    "status:verified",
+    "fused"
+  ]
+}
