+++ src/tilegym/transformers/olmoe/kernel_solutions/olmoe_dual_rms_norm.json
@@ -1,0 +1,23 @@
+{
+  "author": "tilegym-agent",
+  "definition": "olmoe_dual_rms_norm",
+  "description": "cuTile fused in-place Q/K RMSNorm for OLMoE attention.",
+  "name": "olmoe_dual_rms_norm_cutile",
+  "sources": {
+    "path": [
+      "src/tilegym/transformers/olmoe/kernels/dual_rms_norm.py"
+    ]
+  },
+  "spec": {
+    "dependencies": [
+      "torch",
+      "cuda-tile"
+    ],
+    "destination_passing_style": false,
+    "entry_point": "src/tilegym/transformers/olmoe/kernels/dual_rms_norm.py::dual_rms_norm_olmoe_cutile",
+    "language": "cuda-tile",
+    "target_hardware": [
+      "NVIDIA_B200"
+    ]
+  }
+}
