+++ src/tilegym/transformers/qwen3_5/kernel_definitions/qwen3_5_causal_conv1d_update_silu.json
@@ -1,0 +1,76 @@
+{
+  "axes": {
+    "B": {
+      "description": "Supported batch size.",
+      "type": "const",
+      "value": 1
+    },
+    "D": {
+      "description": "Channel dimension.",
+      "type": "var"
+    },
+    "K": {
+      "description": "Convolution kernel size.",
+      "type": "const",
+      "value": 4
+    },
+    "K_state": {
+      "description": "Cached state length.",
+      "type": "const",
+      "value": 3
+    },
+    "S": {
+      "description": "Decode sequence length.",
+      "type": "const",
+      "value": 1
+    }
+  },
+  "constraints": [
+    "K_state == K - 1"
+  ],
+  "description": "Qwen3.5 decode-path depthwise causal conv1d update fused with SiLU. The conv_state input is updated in-place.",
+  "inputs": {
+    "conv_state": {
+      "dtype": "bfloat16",
+      "shape": [
+        "B",
+        "D",
+        "K_state"
+      ]
+    },
+    "hidden_states": {
+      "dtype": "bfloat16",
+      "shape": [
+        "B",
+        "D",
+        "S"
+      ]
+    },
+    "weight": {
+      "dtype": "bfloat16",
+      "shape": [
+        "D",
+        "K"
+      ]
+    }
+  },
+  "name": "qwen3_5_causal_conv1d_update_silu",
+  "op_type": "causal_conv1d_update_silu",
+  "outputs": {
+    "output": {
+      "dtype": "bfloat16",
+      "shape": [
+        "B",
+        "D",
+        "S"
+      ]
+    }
+  },
+  "reference": "# Source: https://github.com/huggingface/transformers/blob/aad13b87ed59f2afcfaebc985f403301887a35fc/src/transformers/models/qwen3_5/modeling_qwen3_5.py#L545-L554\nimport torch\n\ndef run(conv_state, hidden_states, weight):\n    x = hidden_states[:, :, 0]\n    full = torch.cat([conv_state, x.unsqueeze(-1)], dim=-1)\n    dot = (full * weight.unsqueeze(0)).sum(-1)\n    output = dot * torch.sigmoid(dot)\n    conv_state[:, :, 0] = conv_state[:, :, 1]\n    conv_state[:, :, 1] = conv_state[:, :, 2]\n    conv_state[:, :, 2] = x\n    return output.unsqueeze(-1)",
+  "tags": [
+    "model:qwen3.5",
+    "stage:decode",
+    "status:verified",
+    "fused"
+  ]
+}
