+++ src/tilegym/transformers/qwen3_5/kernel_definitions/qwen3_5_sigmoid_mul.json
@@ -1,0 +1,48 @@
+{
+  "axes": {
+    "D": {
+      "description": "Gate/output hidden dimension.",
+      "type": "var"
+    },
+    "N": {
+      "description": "Flattened leading dimensions.",
+      "type": "var"
+    }
+  },
+  "description": "Qwen3.5 fused attention output gate: output = x * sigmoid(gate).",
+  "inputs": {
+    "gate": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    },
+    "x": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    }
+  },
+  "name": "qwen3_5_sigmoid_mul",
+  "op_type": "sigmoid_mul",
+  "outputs": {
+    "output": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    }
+  },
+  "reference": "# Source: https://github.com/huggingface/transformers/blob/aad13b87ed59f2afcfaebc985f403301887a35fc/src/transformers/models/qwen3_5/modeling_qwen3_5.py#L785-L786\nimport torch\n\ndef run(gate, x):\n    return x * torch.sigmoid(gate)",
+  "tags": [
+    "model:qwen3.5",
+    "stage:prefill",
+    "stage:decode",
+    "status:verified",
+    "fused"
+  ]
+}
