+++ src/tilegym/transformers/qwen3_5/kernel_definitions/qwen3_5_silu_and_mul_separate.json
@@ -1,0 +1,48 @@
+{
+  "axes": {
+    "D": {
+      "description": "Intermediate dimension.",
+      "type": "var"
+    },
+    "N": {
+      "description": "Flattened leading dimensions.",
+      "type": "var"
+    }
+  },
+  "description": "Qwen3.5 fused SwiGLU activation for separate gate and up tensors.",
+  "inputs": {
+    "gate": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    },
+    "up": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    }
+  },
+  "name": "qwen3_5_silu_and_mul_separate",
+  "op_type": "silu_and_mul",
+  "outputs": {
+    "output": {
+      "dtype": "bfloat16",
+      "shape": [
+        "N",
+        "D"
+      ]
+    }
+  },
+  "reference": "# Source: https://github.com/huggingface/transformers/blob/aad13b87ed59f2afcfaebc985f403301887a35fc/src/transformers/models/qwen3_5/modeling_qwen3_5.py#L803-L805\nimport torch\n\ndef run(gate, up):\n    return (gate * torch.sigmoid(gate) * up).to(up.dtype)",
+  "tags": [
+    "model:qwen3.5",
+    "stage:prefill",
+    "stage:decode",
+    "status:verified",
+    "fused"
+  ]
+}
