+++ src/tilegym/transformers/qwen3_5/modeling_qwen3_5.py
@@ -19,432 +19,18 @@ gated delta rule linear attention layers.  This module provides:
 import math
 from typing import Optional
 
-import cuda.tile as ct
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from transformers.activations import ACT2FN
 
-from tilegym.ops import silu_and_mul
-
-# Type aliases for cuTile constants
-ConstInt = ct.Constant[int]
-ConstBool = ct.Constant[bool]
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: sigmoid(gate) * x  (for full-attention output gating)
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _sigmoid_mul_kernel(
-    x,  # (N, D)
-    gate,  # (N, D)
-    output,  # (N, D)
-    TILE_D: ConstInt,
-):
-    bid = ct.bid(0)
-    offs = ct.arange(TILE_D, dtype=ct.int32)
-
-    xv = ct.astype(ct.gather(x, (bid, offs), check_bounds=True), ct.float32)
-    gv = ct.astype(ct.gather(gate, (bid, offs), check_bounds=True), ct.float32)
-
-    sig = ct.truediv(1.0, ct.add(1.0, ct.exp(ct.negative(gv))))
-    result = ct.astype(xv * sig, output.dtype)
-    ct.scatter(output, (bid, offs), result, check_bounds=True)
-
-
-def sigmoid_mul_cutile(x: torch.Tensor, gate: torch.Tensor) -> torch.Tensor:
-    """Compute x * sigmoid(gate) using a fused cuTile kernel."""
-    orig_shape = x.shape
-    D = orig_shape[-1]
-    x_flat = x.contiguous().view(-1, D)
-    gate_flat = gate.contiguous().view(-1, D)
-    N = x_flat.shape[0]
-    output = torch.empty_like(x_flat)
-    TILE_D = 1 << (D - 1).bit_length()  # next power of 2
-    ct.launch(torch.cuda.current_stream(), (N,), _sigmoid_mul_kernel, (x_flat, gate_flat, output, TILE_D))
-    return output.view(orig_shape)
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: fused gate preprocessing for gated delta rule
-#   beta = sigmoid(b)
-#   g = -exp(A_log) * softplus(a + dt_bias)
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _gdr_preprocess_kernel(
-    b_in,  # (N, H)
-    a_in,  # (N, H)
-    A_log,  # (H,)
-    dt_bias,  # (H,)
-    beta_out,  # (N, H)
-    g_out,  # (N, H)
-    TILE_H: ConstInt,
-):
-    bid = ct.bid(0)  # row index in [0, N)
-    offs = ct.arange(TILE_H, dtype=ct.int32)
-
-    b = ct.astype(ct.gather(b_in, (bid, offs), check_bounds=True), ct.float32)
-    a = ct.astype(ct.gather(a_in, (bid, offs), check_bounds=True), ct.float32)
-    a_log = ct.astype(ct.gather(A_log, (offs,), check_bounds=True), ct.float32)
-    dt_b = ct.astype(ct.gather(dt_bias, (offs,), check_bounds=True), ct.float32)
-
-    # beta = sigmoid(b)
-    beta = ct.truediv(1.0, ct.add(1.0, ct.exp(ct.negative(b))))
-
-    # g = -exp(A_log) * softplus(a + dt_bias)
-    # softplus(x) = log(1 + exp(x))
-    sp_arg = a + dt_b
-    sp = ct.log(ct.add(1.0, ct.exp(sp_arg)))
-    g = ct.negative(ct.exp(a_log) * sp)
-
-    ct.scatter(beta_out, (bid, offs), ct.astype(beta, beta_out.dtype), check_bounds=True)
-    ct.scatter(g_out, (bid, offs), ct.astype(g, g_out.dtype), check_bounds=True)
-
-
-def gdr_preprocess_cutile(
-    b: torch.Tensor, a: torch.Tensor, A_log: torch.Tensor, dt_bias: torch.Tensor
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Fused computation of beta=sigmoid(b) and g=-exp(A_log)*softplus(a+dt_bias)."""
-    orig_shape = b.shape
-    H = orig_shape[-1]
-    b_flat = b.contiguous().view(-1, H)
-    a_flat = a.contiguous().view(-1, H)
-    N = b_flat.shape[0]
-    beta = torch.empty_like(b_flat)
-    g = torch.empty(N, H, dtype=torch.float32, device=b.device)
-    TILE_H = 1 << (H - 1).bit_length()
-    ct.launch(
-        torch.cuda.current_stream(),
-        (N,),
-        _gdr_preprocess_kernel,
-        (b_flat, a_flat, A_log, dt_bias, beta, g, TILE_H),
-    )
-    return beta.view(orig_shape), g.view(orig_shape)
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: fused RMSNormGated = weight * norm(x) * silu(gate)
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _rms_norm_gated_silu_kernel(
-    hidden_states,  # (N, D)
-    gate,  # (N, D)
-    weight,  # (D,)
-    output,  # (N, D)
-    eps: float,
-    D: ConstInt,
-    TILE_D: ConstInt,
-):
-    bid = ct.bid(0)
-    offs = ct.arange(TILE_D, dtype=ct.int32)
-
-    h = ct.astype(ct.gather(hidden_states, (bid, offs), padding_value=0.0, check_bounds=True), ct.float32)
-    g = ct.astype(ct.gather(gate, (bid, offs), padding_value=0.0, check_bounds=True), ct.float32)
-    w = ct.astype(ct.gather(weight, (offs,), padding_value=0.0, check_bounds=True), ct.float32)
-
-    # RMSNorm: h / sqrt(mean(h^2) + eps)
-    variance = ct.sum(h * h) * ct.truediv(1.0, D)
-    normed = h * ct.rsqrt(variance + eps)
-
-    # Apply weight and SiLU gate: weight * normed * silu(gate)
-    sig_g = ct.truediv(1.0, ct.add(1.0, ct.exp(ct.negative(g))))
-    silu_g = g * sig_g
-    result = ct.astype(w * normed * silu_g, output.dtype)
-
-    ct.scatter(output, (bid, offs), result, check_bounds=True)
-
-
-def rms_norm_gated_cutile(
-    hidden_states: torch.Tensor,
-    gate: torch.Tensor,
-    weight: torch.Tensor,
-    eps: float,
-) -> torch.Tensor:
-    """Fused RMSNorm + SiLU gating as cuTile kernel."""
-    D = hidden_states.shape[-1]
-    h_flat = hidden_states.contiguous().view(-1, D)
-    g_flat = gate.contiguous().view(-1, D)
-    N = h_flat.shape[0]
-    output = torch.empty_like(h_flat)
-    TILE_D = 1 << (D - 1).bit_length()
-    ct.launch(
-        torch.cuda.current_stream(),
-        (N,),
-        _rms_norm_gated_silu_kernel,
-        (h_flat, g_flat, weight, output, eps, D, TILE_D),
-    )
-    return output.view(hidden_states.shape)
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: fused residual add + Gemma-style RMSNorm
-#   Computes: sum = residual + x; normed = (1+w) * sum / sqrt(mean(sum^2) + eps)
-#   Outputs both sum and normed (for residual chain + next block input)
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _residual_add_rms_norm_kernel(
-    residual,  # (N, D)
-    x,  # (N, D)
-    weight,  # (D,)
-    sum_out,  # (N, D) — residual + x
-    normed_out,  # (N, D) — rms_norm(sum)
-    eps: float,
-    offset: float,
-    D: ConstInt,
-    TILE_D: ConstInt,
-):
-    bid = ct.bid(0)
-    offs = ct.arange(TILE_D, dtype=ct.int32)
-
-    r = ct.astype(ct.gather(residual, (bid, offs), padding_value=0.0, check_bounds=True), ct.float32)
-    h = ct.astype(ct.gather(x, (bid, offs), padding_value=0.0, check_bounds=True), ct.float32)
-    w = ct.astype(ct.gather(weight, (offs,), padding_value=0.0, check_bounds=True), ct.float32)
-
-    # Residual add
-    s = r + h
-
-    # Gemma-style RMSNorm: (offset + weight) * s / sqrt(mean(s^2) + eps)
-    variance = ct.sum(s * s) * ct.truediv(1.0, D)
-    normed = s * ct.rsqrt(variance + eps) * (offset + w)
-
-    ct.scatter(sum_out, (bid, offs), ct.astype(s, sum_out.dtype), check_bounds=True)
-    ct.scatter(normed_out, (bid, offs), ct.astype(normed, normed_out.dtype), check_bounds=True)
-
-
-def residual_add_rms_norm_cutile(
-    residual: torch.Tensor,
-    x: torch.Tensor,
-    weight: torch.Tensor,
-    eps: float,
-    offset: float = 1.0,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Fused residual add + Gemma-style RMSNorm. Returns (sum, normed)."""
-    D = residual.shape[-1]
-    r_flat = residual.contiguous().view(-1, D)
-    x_flat = x.contiguous().view(-1, D)
-    N = r_flat.shape[0]
-    sum_out = torch.empty_like(r_flat)
-    normed_out = torch.empty_like(r_flat)
-    TILE_D = 1 << (D - 1).bit_length()
-    ct.launch(
-        torch.cuda.current_stream(),
-        (N,),
-        _residual_add_rms_norm_kernel,
-        (r_flat, x_flat, weight, sum_out, normed_out, eps, offset, D, TILE_D),
-    )
-    return sum_out.view(residual.shape), normed_out.view(residual.shape)
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: fused causal conv1d update + SiLU (decode path)
-#   Fuses: cat + depthwise conv1d + silu + state update
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _causal_conv1d_update_silu_kernel(
-    x,  # (D,) — input for current time step
-    conv_state,  # (D, 3) — state buffer (kernel_size - 1)
-    weight,  # (D, 4) — conv weights
-    output,  # (D,) — output
-    BLOCK_D: ConstInt,
-):
-    bid = ct.bid(0)
-    d_start = bid * BLOCK_D
-    offs = ct.arange(BLOCK_D, dtype=ct.int32)
-    d_idx = d_start + offs
-
-    # Load state (3 values per channel) and input
-    s0 = ct.astype(ct.gather(conv_state, (d_idx, 0), check_bounds=True), ct.float32)
-    s1 = ct.astype(ct.gather(conv_state, (d_idx, 1), check_bounds=True), ct.float32)
-    s2 = ct.astype(ct.gather(conv_state, (d_idx, 2), check_bounds=True), ct.float32)
-    xv = ct.astype(ct.gather(x, (d_idx,), check_bounds=True), ct.float32)
-
-    # Load weights
-    w0 = ct.astype(ct.gather(weight, (d_idx, 0), check_bounds=True), ct.float32)
-    w1 = ct.astype(ct.gather(weight, (d_idx, 1), check_bounds=True), ct.float32)
-    w2 = ct.astype(ct.gather(weight, (d_idx, 2), check_bounds=True), ct.float32)
-    w3 = ct.astype(ct.gather(weight, (d_idx, 3), check_bounds=True), ct.float32)
-
-    # Dot product: conv1d depthwise
-    dot = s0 * w0 + s1 * w1 + s2 * w2 + xv * w3
-
-    # SiLU activation
-    sig = ct.truediv(1.0, ct.add(1.0, ct.exp(ct.negative(dot))))
-    result = dot * sig
-
-    # Store output
-    ct.scatter(output, (d_idx,), ct.astype(result, output.dtype), check_bounds=True)
-
-    # Update conv state: shift left by 1, append new input
-    ct.scatter(conv_state, (d_idx, 0), ct.astype(s1, conv_state.dtype), check_bounds=True)
-    ct.scatter(conv_state, (d_idx, 1), ct.astype(s2, conv_state.dtype), check_bounds=True)
-    ct.scatter(conv_state, (d_idx, 2), ct.astype(xv, conv_state.dtype), check_bounds=True)
-
-
-def causal_conv1d_update_silu_cutile(
-    hidden_states: torch.Tensor,
-    conv_state: torch.Tensor,
-    weight: torch.Tensor,
-    bias=None,
-    activation=None,
-) -> torch.Tensor:
-    """Fused causal conv1d update + SiLU for decode path (seq_len=1).
-
-    Args:
-        hidden_states: (B, D, 1) input
-        conv_state: (B, D, kernel_size-1) state buffer, updated in-place
-        weight: (D, kernel_size) conv weights
-        bias: ignored (Qwen3.5 conv has no bias)
-        activation: ignored (always SiLU)
-    """
-    B, D, seq_len = hidden_states.shape
-    assert seq_len == 1, "causal_conv1d_update_silu_cutile only supports seq_len=1"
-    assert B == 1, "causal_conv1d_update_silu_cutile only supports B=1 currently"
-
-    x = hidden_states.squeeze(0).squeeze(-1).contiguous()  # (D,)
-    w = weight.contiguous()  # (D, 4)
-    output = torch.empty(D, dtype=hidden_states.dtype, device=hidden_states.device)
-    cs = conv_state.squeeze(0)  # (D, 3)
-
-    BLOCK_D = 256
-    grid = ((D + BLOCK_D - 1) // BLOCK_D,)
-    ct.launch(
-        torch.cuda.current_stream(),
-        grid,
-        _causal_conv1d_update_silu_kernel,
-        (x, cs, w, output, BLOCK_D),
-    )
-    return output.unsqueeze(0).unsqueeze(-1)  # (1, D, 1)
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: SiLU(gate) * up — takes separate inputs (no cat needed)
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _silu_and_mul_separate_kernel(
-    gate,  # (N, D)
-    up,  # (N, D)
-    output,  # (N, D)
-    TILE_D: ConstInt,
-):
-    bid = ct.bid(0)
-    offs = ct.arange(TILE_D, dtype=ct.int32)
-
-    g = ct.astype(ct.gather(gate, (bid, offs), check_bounds=True), ct.float32)
-    u = ct.astype(ct.gather(up, (bid, offs), check_bounds=True), ct.float32)
-
-    # SiLU(gate) * up = gate * sigmoid(gate) * up
-    sig = ct.truediv(1.0, ct.add(1.0, ct.exp(ct.negative(g))))
-    result = ct.astype(g * sig * u, output.dtype)
-    ct.scatter(output, (bid, offs), result, check_bounds=True)
-
-
-def silu_and_mul_separate_cutile(gate: torch.Tensor, up: torch.Tensor) -> torch.Tensor:
-    """Compute silu(gate) * up without concatenation, using a fused cuTile kernel."""
-    orig_shape = gate.shape
-    D = orig_shape[-1]
-    gate_flat = gate.contiguous().view(-1, D)
-    up_flat = up.contiguous().view(-1, D)
-    N = gate_flat.shape[0]
-    output = torch.empty_like(gate_flat)
-    TILE_D = 1 << (D - 1).bit_length()
-    ct.launch(torch.cuda.current_stream(), (N,), _silu_and_mul_separate_kernel, (gate_flat, up_flat, output, TILE_D))
-    return output.view(orig_shape)
-
-
-# ──────────────────────────────────────────────────────────────────────
-# cuTile kernel: fused depthwise conv1d + SiLU for prefill path
-#   Replaces: F.silu(self.conv1d(mixed_qkv)[:, :, :seq_len])
-# ──────────────────────────────────────────────────────────────────────
-
-
-@ct.kernel
-def _causal_conv1d_prefill_silu_kernel(
-    x,  # (D, T) — input channels × time
-    weight,  # (D, K) — conv weights per channel, K=kernel_size
-    output,  # (D, T) — output
-    T: ConstInt,
-    BLOCK_T: ConstInt,
-):
-    """Each block handles one channel across a tile of time steps."""
-    bid_d = ct.bid(0)  # channel index
-    bid_t = ct.bid(1)  # time tile index
-    t_start = bid_t * BLOCK_T
-    offs = ct.arange(BLOCK_T, dtype=ct.int32)
-    t_idx = t_start + offs
-
-    # Load 4 weights for this channel (kernel_size=4)
-    w0 = ct.astype(ct.gather(weight, (bid_d, 0), check_bounds=True), ct.float32)
-    w1 = ct.astype(ct.gather(weight, (bid_d, 1), check_bounds=True), ct.float32)
-    w2 = ct.astype(ct.gather(weight, (bid_d, 2), check_bounds=True), ct.float32)
-    w3 = ct.astype(ct.gather(weight, (bid_d, 3), check_bounds=True), ct.float32)
-
-    # Load input values: x[d, t], x[d, t+1], x[d, t+2], x[d, t+3]
-    # The conv1d with padding=3 means output[t] = sum(x[t:t+4] * w)
-    v0 = ct.astype(ct.gather(x, (bid_d, t_idx), padding_value=0.0, check_bounds=True), ct.float32)
-    v1 = ct.astype(ct.gather(x, (bid_d, t_idx + 1), padding_value=0.0, check_bounds=True), ct.float32)
-    v2 = ct.astype(ct.gather(x, (bid_d, t_idx + 2), padding_value=0.0, check_bounds=True), ct.float32)
-    v3 = ct.astype(ct.gather(x, (bid_d, t_idx + 3), padding_value=0.0, check_bounds=True), ct.float32)
-
-    # Depthwise conv: dot product per time step
-    dot = v0 * w0 + v1 * w1 + v2 * w2 + v3 * w3
-
-    # SiLU activation
-    sig = ct.truediv(1.0, ct.add(1.0, ct.exp(ct.negative(dot))))
-    result = dot * sig
-
-    ct.scatter(output, (bid_d, t_idx), ct.astype(result, output.dtype), check_bounds=True)
-
-
-def causal_conv1d_prefill_silu_cutile(
-    x: torch.Tensor,
-    weight: torch.Tensor,
-    seq_len: int,
-) -> torch.Tensor:
-    """Fused causal depthwise conv1d + SiLU for prefill path.
-
-    Args:
-        x: (B, D, T_padded) input after nn.Conv1d padding
-        weight: (D, kernel_size) conv weights
-        seq_len: actual sequence length to slice output
-    Returns:
-        (B, D, seq_len) output after conv1d + SiLU
-    """
-    B, D, T_padded = x.shape
-    assert B == 1, "causal_conv1d_prefill_silu only supports B=1"
-
-    # The nn.Conv1d with padding=kernel_size-1=3 pads on BOTH sides.
-    # For causal conv, we want output[:, :, :seq_len].
-    # The padded input has T_padded = seq_len + 3 (padding on right).
-    # conv1d output at position t = sum(x[t:t+4] * w) for t in [0, T_padded-3)
-    # We want the first seq_len positions.
-
-    x_2d = x.squeeze(0).contiguous()  # (D, T_padded)
-    w = weight.contiguous()  # (D, 4)
-    output = torch.empty(D, seq_len, dtype=x.dtype, device=x.device)
-
-    BLOCK_T = 256
-    grid = (D, (seq_len + BLOCK_T - 1) // BLOCK_T)
-    ct.launch(
-        torch.cuda.current_stream(),
-        grid,
-        _causal_conv1d_prefill_silu_kernel,
-        (x_2d, w, output, seq_len, BLOCK_T),
-    )
-    return output.unsqueeze(0)  # (1, D, seq_len)
-
+from tilegym.transformers.qwen3_5.kernels.causal_conv1d_prefill_silu import causal_conv1d_prefill_silu_cutile
+from tilegym.transformers.qwen3_5.kernels.causal_conv1d_update_silu import causal_conv1d_update_silu_cutile
+from tilegym.transformers.qwen3_5.kernels.gdr_preprocess import gdr_preprocess_cutile
+from tilegym.transformers.qwen3_5.kernels.residual_add_rms_norm import residual_add_rms_norm_cutile
+from tilegym.transformers.qwen3_5.kernels.rms_norm_gated import rms_norm_gated_cutile
+from tilegym.transformers.qwen3_5.kernels.sigmoid_mul import sigmoid_mul_cutile
+from tilegym.transformers.qwen3_5.kernels.silu_and_mul_separate import silu_and_mul_separate_cutile
 
 # ──────────────────────────────────────────────────────────────────────
 # Replacement Qwen3_5RMSNormGated using cuTile
